fixed delete functions, query owner as well

+++ src/contexts/AlbumContext.js
@@ -71,8 +71,11 @@ const AlbumContextProvider = (props) => {
 
 					// delete image from firestore
 					imageDoc.ref.delete();
-					// query all images with matching path as imageDoc
-					const snapshot = await db.collection('images').where('path', '==', imageDoc.data().path).get();
+					// query all images with matching path and owner as imageDoc
+					const snapshot = await db.collection('images')
+						.where('owner', '==', album.owner)
+						.where('path', '==', imageDoc.data().path)
+						.get();
 					const snapshotImgs = []
 					snapshot.forEach(doc => {
 						if (snapshotImgs.includes(doc.data())) {
@@ -83,10 +86,9 @@ const AlbumContextProvider = (props) => {
 							...doc.data()
 						})
 					});
-
 					// if no image with the queried path exist in firestore
 					if (snapshotImgs.length === 0) {
-						// delete the image from storage
+						// delete image from storage
 						storage.ref(imageDoc.data().path)
 						.delete()
 						.then(() => {
